+++ src/background/api.ts
@@ -2,6 +2,7 @@ import { StaticConfig } from '../JsonSettings.interface';
 import { AppTemplate, UserSettings } from '../shared.interfaces';
 import { ApplicationConfiguration } from '../YamlSettings.interface';
 import { Channel, REPLY_BY_CHANNEL } from './constants';
+import { Version } from './init';
 import { execPrinter, fromPackageRoot, getEnviroment, runPwshScript } from './utils';
 import { exec } from 'child_process';
 import { app, BrowserWindow, dialog, ipcMain } from 'electron';
@@ -23,13 +24,8 @@ const tryRunAhkShortcuts = () => {
 
 export const loadBackgroundApi = (mainWindow: BrowserWindow) => {
   ipcMain.on(Channel.ENABLE_AUTOSTART, (_event) => {
-    switch (getEnviroment()) {
-      case 'installed':
-        return runPwshScript('autostart_on.ps1', `-VersionPath "${fromPackageRoot('../version')}"`);
-      case 'packaged':
-        return runPwshScript('autostart_on.ps1', `-ExeRoute "${fromPackageRoot('/komorebi.exe')}"`);
-      default:
-        return;
+    if (getEnviroment() !== 'dev') {
+      runPwshScript('autostart_on.ps1', `-VersionPath "${Version.pathToSave}"`);
     }
   });