+++ src/background/init.ts
@@ -1,26 +1,28 @@
-import { fromPackageRoot, getEnviroment, runPwshScript } from './utils';
+import { fromPackageRoot, runPwshScript } from './utils';
 import { app } from 'electron';
 import { copyFileSync, existsSync, readFileSync, writeFileSync } from 'fs';
+import { ensureFileSync } from 'fs-extra';
 import path from 'path';
 
-interface Version {
+export interface Version {
   version: string;
   path: string;
 }
 
-class Version {
-  static path = getEnviroment() === 'installed' ? fromPackageRoot('../version') : fromPackageRoot('/version');
+export class Version {
+  static pathToSave = path.join(app.getPath('userData'), 'version');
 
   static getVersionFile(): Version | undefined {
-    if (!existsSync(Version.path)) {
+    if (!existsSync(Version.pathToSave)) {
       return;
     }
-    return JSON.parse(readFileSync(Version.path, 'utf-8'));
+    return JSON.parse(readFileSync(Version.pathToSave, 'utf-8'));
   }
 
   static updateVersionFile(): Version {
     const data: Version = { version: app.getVersion(), path: fromPackageRoot() };
-    writeFileSync(Version.path, JSON.stringify(data));
+    ensureFileSync(Version.pathToSave);
+    writeFileSync(Version.pathToSave, JSON.stringify(data));
     return data;
   }
 }
@@ -37,16 +39,18 @@ export const init = async () => {
     return;
   }
 
+  extract('komorebi.exe');
+  extract('komorebic.exe');
+
   const oldVersion = Version.getVersionFile();
   const currentVersion = Version.updateVersionFile();
 
-  // execute install and update tasks
+  // fresh install or update
   if (!oldVersion || oldVersion.version != currentVersion.version) {
-    extract('komorebi.exe');
-    extract('komorebic.exe');
     await runPwshScript('update_path.ps1', `-AppPath "${currentVersion.path}\\"`);
+
     if (oldVersion) {
-      await runPwshScript('update_path.ps1', `-AppPath "${oldVersion.path}\\" -Delete $true`);
+      await runPwshScript('update_path.ps1', `-AppPath "${oldVersion.path}\\" -Delete true`);
       await runPwshScript('force_stop.ps1', `-ExeRoute "${fromPackageRoot('komorebi.exe')}"`);
     }
   }