+++ static/pwsh/update_path.ps1
@@ -1,12 +1,13 @@
 param (
   [string]$AppPath,
-  [bool]$Delete
+  [string]$Delete
 )
 
 $existingPath = [System.Environment]::GetEnvironmentVariable("PATH", [System.EnvironmentVariableTarget]::User)
 
-if ($Delete) {
-  $newPath = $existingPath -replace [regex]::Escape($AppPath) + ';', ''
+if ($Delete -eq "true") {
+  $pathToDelete = [regex]::Escape($AppPath)
+  $newPath = ($existingPath -split ';' | Where-Object {$_ -notmatch "^$pathToDelete$"}) -join ';'
   [Environment]::SetEnvironmentVariable("PATH", $newPath, [System.EnvironmentVariableTarget]::User)
 } else {
   if ($existingPath -split ';' -notcontains $AppPath) {