+++ src/app/app.tsx
@@ -1,8 +1,10 @@
 import { Header } from './components/header';
 import { Navigation } from './components/navigation';
+import { Suspense } from 'react';
 
 import { AppsConfiguration } from './modules/appsConfigurations/infra/infra';
 import { General } from './modules/general/main/infra';
+import { StylesView } from './modules/general/visuals/infra';
 import { Information } from './modules/information/infrastructure';
 import { Monitors } from './modules/monitors/main/infra';
 import { Shortcuts } from './modules/shortcuts/infrastructure';
@@ -15,6 +17,7 @@ import { Route } from './modules/shared/domain/routes';
 const ComponentByRout: Record<Route, React.JSXElementConstructor<any>> = {
   [Route.GENERAL]: General,
   [Route.MONITORS]: Monitors,
+  [Route.STYLES]: StylesView,
   [Route.SHORTCUTS]: Shortcuts,
   [Route.SPECIFIT_APPS]: AppsConfiguration,
   [Route.INFO]: Information,
@@ -24,11 +27,15 @@ export function App() {
   let route = useAppSelector(RootSelectors.route);
   let Component = ComponentByRout[route];
 
-  return <>
-    <Navigation />
-    <Header />
-    <div className="content">
-      <Component />
-    </div>
-  </>;
+  return (
+    <>
+      <Navigation />
+      <Header />
+      <div className="content">
+        <Suspense fallback={<div>Loading...</div>}>
+          <Component />
+        </Suspense>
+      </div>
+    </>
+  );
 }