+++ src/app/components/SettingsBox/index.tsx
@@ -1,3 +1,5 @@
+import { ConfigProvider } from 'antd';
+
 import cs from './index.module.css';
 
 interface Props {
@@ -5,27 +7,46 @@ interface Props {
 }
 
 export const SettingsGroup = ({ children }: Props) => {
-  return <div className={cs.group}>
-    {children}
-  </div>;
+  return <div className={cs.group}>{children}</div>;
 };
 
 interface SubGroupProps {
   children: React.ReactNode;
   label: React.ReactNode;
+  disableOptions?: boolean;
 }
 
-export const SettingsSubGroup = ({ children, label }: SubGroupProps) => {
-  return <div>
-    <div className={cs.subtitle}>{label}</div>
-    <div className={cs.subgroup}>
-      {children}
+export const SettingsSubGroup = ({ children, label, disableOptions }: SubGroupProps) => {
+  return (
+    <div>
+      <div className={cs.subtitle}>{label}</div>
+      <ConfigProvider componentDisabled={disableOptions}>
+        <div className={cs.subgroup}>{children}</div>
+      </ConfigProvider>
     </div>
-  </div>;
+  );
 };
 
-export const SettingsOption = ({ children }: Props) => {
-  return <div className={cs.box}>
-    {children}
-  </div>;
+type OptionProps =
+  | {
+    children: React.ReactNode;
+  }
+  | {
+    label: React.ReactNode;
+    trigger: React.ReactNode;
+  };
+
+export const SettingsOption = (props: OptionProps) => {
+  return (
+    <div className={cs.box}>
+      {'children' in props ? (
+        props.children
+      ) : (
+        <>
+          <span>{props.label}</span>
+          {props.trigger}
+        </>
+      )}
+    </div>
+  );
 };