+++ src/app/modules/general/main/infra/index.tsx
@@ -1,19 +1,15 @@
 import {
   SettingsGroup,
   SettingsOption,
-  SettingsSubGroup,
 } from '../../../../components/SettingsBox';
 import { ContainerBehaviors } from './ContainerBehaviours';
 import { FocusBehaviours } from './FocusBehaviours';
 import { GlobalPaddings } from './GlobalPaddings';
 import { OthersConfigs } from './Others';
-import { InputNumber, Select, Switch } from 'antd';
+import { Switch } from 'antd';
 import { useEffect, useState } from 'react';
-import { useSelector } from 'react-redux';
 
 import { AnimationsSettings } from '../../animations/infra';
-import { BorderSettings } from '../../border/infra';
-import { ContainerTopBarSettings } from '../../containerTopBar/infra';
 
 export function General() {
   const [autostart, setAutostart] = useState(false);
@@ -42,14 +38,12 @@ export function General() {
       </SettingsGroup>
 
       <SettingsGroup>
-        <BorderSettings />
         <FocusBehaviours />
       </SettingsGroup>
 
       <ContainerBehaviors />
       <AnimationsSettings />
       <GlobalPaddings/>
-      <ContainerTopBarSettings />
 
       <OthersConfigs />
     </>