+++ src/app/modules/general/popups/domain.ts
@@ -1,0 +1,11 @@
+export interface PopupState {
+  enable: boolean;
+  x: number | null;
+  y: number | null;
+  height: number;
+  width: number;
+  textColor: string;
+  background: string;
+  borderColor: string;
+  borderWidth: number;
+};