+++ src/app/modules/general/visuals/infra.tsx
@@ -1,0 +1,17 @@
+import { SettingsGroup } from '../../../components/SettingsBox';
+
+import { BorderSettings } from '../border/infra';
+import { ContainerTopBarSettings } from '../containerTopBar/infra';
+import { PopupsSettings } from '../popups/infra';
+
+export const StylesView = () => {
+  return (
+    <>
+      <PopupsSettings />
+      <ContainerTopBarSettings />
+      <SettingsGroup>
+        <BorderSettings />
+      </SettingsGroup>
+    </>
+  );
+};