+++ src/app/modules/shared/domain/routes.ts
@@ -1,6 +1,7 @@
 export enum Route {
   GENERAL = 'general',
   MONITORS = 'monitors',
+  STYLES = 'styles',
   SPECIFIT_APPS = 'specifit_apps',
   SHORTCUTS = 'shortcuts',
   INFO = 'info',
@@ -9,6 +10,7 @@ export enum Route {
 export const RouteLabels: Record<Route, string> = {
   [Route.GENERAL]: 'General',
   [Route.MONITORS]: 'Monitors',
+  [Route.STYLES]: 'Visuals',
   [Route.SHORTCUTS]: 'Shortcuts',
   [Route.SPECIFIT_APPS]: 'Apps Configurations',
   [Route.INFO]: 'Information',
@@ -17,6 +19,7 @@ export const RouteLabels: Record<Route, string> = {
 export const RouteExtraInfo: Record<Route, string | null> = {
   [Route.GENERAL]: null,
   [Route.MONITORS]: null,
+  [Route.STYLES]: null,
   [Route.SHORTCUTS]: null,
   [Route.SPECIFIT_APPS]: `
     Komorebi-UI use only one identifier per app (first match found) so the order in how are specificated is important,
@@ -28,6 +31,7 @@ export const RouteExtraInfo: Record<Route, string | null> = {
 export const RouteIcons: Record<Route, string> = {
   [Route.GENERAL]: '⚙️',
   [Route.MONITORS]: '🖥️',
+  [Route.STYLES]: '🖼️',
   [Route.SHORTCUTS]: '🔡',
   [Route.SPECIFIT_APPS]: '🅰️',
   [Route.INFO]: '🛈',