+++ changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### added
+- add popups on actions 🦀
+
 ## [1.0.0-beta.8]
 ### added
 - add more templates