+++ changelog.md
@@ -3,6 +3,10 @@
 ## [Unreleased]
 ### added
 - add popups on actions 🦀
+- now can switch from installed and packaged and should work as the same
+
+### fixed
+- fix no removing old path
 
 ## [1.0.0-beta.8]
 ### added